requirement.txt
streamlit
python-dotenv
langchain
langchain_google_genai
faiss-cpu
langchain_groq
langchain_community
groq
beautifulsoup4